requirements.txt
PyQt6>=6.6.0
yt-dlp>=2024.1.0
Pillow>=10.0.0
